# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: VITRAS APS API
  version: "1.0.0"
  description: >
    API do sistema VITRAS APS para Atenção Primária à Saúde.
    Autenticação via JWT (Bearer) ou cookie httpOnly (vitras_access).
    Todos os endpoints autenticados exigem CSRF token (X-CSRF-Token) quando usando cookie.
    CPF e CNS em listagens são sempre mascarados (últimos 2 dígitos visíveis).
  contact:
    name: Equipe VITRAS
servers:
  - url: https://api.vitras.com.br
    description: Produção
  - url: http://localhost:3000
    description: Desenvolvimento local
security:
  - cookieAuth: []
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT obtido via /auth/login. Validade configurada (padrão: 15m)."
    cookieAuth:
      type: apiKey
      in: cookie
      name: vitras_access
      description: "Cookie httpOnly definido em /auth/login. Requer X-CSRF-Token para mutações."
  schemas:
...
    Household:
      type: object
      properties:
        id: { type: string, format: uuid }
        patientId: { type: string }
        patientIds: { type: array, items: { type: string } }
        teamId: { type: string }
        logradouro: { type: string }
        numero: { type: string }
        bairro: { type: string }
        cep: { type: string }
        uf: { type: string }
        municipioIbge: { type: string }
        tipoImovel: { type: string }
        tipoDomicilio: { type: string }
        localizacao: { type: string, enum: [URBANA, RURAL, PERIURBANA] }
        abastecimentoAgua: { type: string }
        esgotamento: { type: string }
        destinacaoLixo: { type: string }
        energiaEletrica: { type: boolean }
        microarea: { type: string }
        familias: { type: array, items: { type: object } }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
paths:
...
  /households:
    get:
      summary: Listar domicílios de um paciente
      description: Requer patientId como query param. Valida que paciente pertence à equipe do usuário.
      tags: [Households]
      parameters:
        - name: patientId
          in: query
          required: true
          schema: { type: string, format: uuid }
          description: ID do paciente
      responses:
        "200":
          description: Lista de domicílios do paciente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Household"
        "400":
          description: patientId obrigatório
        "403":
          description: Sem permissão para visualizar domicílios desta equipe
        "404":
          description: Paciente não encontrado

    post:
      summary: Criar Cadastro Domiciliar e Territorial
      description: >
        Cria novo domicílio ou reutiliza existente com mesmo endereço na mesma equipe.
        Vincula paciente ao domicílio.
      tags: [Households]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patientId]
              properties:
                patientId: { type: string, format: uuid }
                tipoImovel: { type: string }
                tipoDomicilio: { type: string }
                localizacao: { type: string, enum: [URBANA, RURAL, PERIURBANA] }
                logradouro: { type: string }
                numero: { type: string }
                semNumero: { type: boolean }
                complemento: { type: string }
                bairro: { type: string }
                cep: { type: string }
                uf: { type: string }
                municipioIbge: { type: string }
                microarea: { type: string }
                abastecimentoAgua: { type: string }
                tratamentoAgua: { type: string }
                esgotamento: { type: string }
                destinacaoLixo: { type: string }
                energiaEletrica: { type: boolean }
                origemEnergiaEletrica: { type: string }
                materialPredominanteParedes: { type: string }
                situacaoMoradiaPosseTerra: { type: string }
                tipoAcesso: { type: string }
                numMoradores: { type: integer, minimum: 0 }
                numComodos: { type: integer, minimum: 0 }
                animaisNoDomicilio: { type: boolean }
                tiposAnimais: { type: array, items: { type: string } }
                familias: { type: array, maxItems: 20, items: { type: object } }
                termoRecusa: { type: boolean }
                dataCadastro: { type: string, format: date }
      responses:
        "201":
          description: Domicílio criado (ou reutilizado e paciente vinculado)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Household"
        "400":
          description: Dados inválidos
        "403":
          description: Sem permissão para criar domicílio nesta equipe
        "404":
          description: Paciente não encontrado
  /households/{id}:
    get:
      summary: Detalhe de domicílio por ID
      description: Valida que domicílio pertence à equipe do usuário (exceto break_glass_admin).
      tags: [Households]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Dados do domicílio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Household"
        "401":
          description: Não autenticado
        "403":
          description: Sem permissão para visualizar este domicílio
        "404":
          description: Domicílio não encontrado

    patch:
      summary: Atualizar Cadastro Domiciliar e Territorial
      tags: [Households]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Campos atualizáveis do domicílio. Todos opcionais (PATCH parcial).
              properties:
                tipoImovel: { type: string }
                tipoDomicilio: { type: string }
                localizacao: { type: string }
                abastecimentoAgua: { type: string }
                esgotamento: { type: string }
                microarea: { type: string }
                familias: { type: array, maxItems: 20, items: { type: object } }
      responses:
        "200":
          description: Domicílio atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Household"
        "400":
          description: Dados inválidos
        "403":
          description: Sem permissão
        "404":
          description: Domicílio não encontrado
        "409":
          description: Conflito de concorrência — registro foi alterado por outro usuário

  # ─── ACS ──────────────────────────────────────────────────────────────────
...